refactor(project-details): rename ProjectInformationProvider.tsx, use barrel imports for better readability.
