+++ client/src/features/project-details/components/index.ts
@@ -1,0 +1,11 @@
+import CoreContributors from "./CoreContributors";
+import ProjectImages from "./ProjectImages";
+import ProjectInformationCard from "./ProjectInformation";
+import ProjectInformationProvider from "./ProjectInformationProvider";
+
+export {
+    CoreContributors,
+    ProjectImages,
+    ProjectInformationCard,
+    ProjectInformationProvider
+};
